working properly with web3modal
