+++ frontend/src/app/layout.js
@@ -1,17 +1,46 @@
-import "./globals.css";
+"use client";
+import "@/app/globals.css";
+import Header from "@/components/header";
 import { Inter } from "next/font/google";
+
+import {
+  EthereumClient,
+  w3mConnectors,
+  w3mProvider,
+} from "@web3modal/ethereum";
+import { Web3Modal } from "@web3modal/react";
+import { configureChains, createConfig, WagmiConfig } from "wagmi";
+import { sepolia } from "wagmi/chains";
+
+const chains = [sepolia];
+const projectId = "c982d972956527324406e9e235eb10ed";
+
+const { publicClient } = configureChains(chains, [w3mProvider({ projectId })]);
+const wagmiConfig = createConfig({
+  autoConnect: true,
+  connectors: w3mConnectors({ projectId, chains }),
+  publicClient,
+});
+const ethereumClient = new EthereumClient(wagmiConfig, chains);
 
 const inter = Inter({ subsets: ["latin"] });
 
-export const metadata = {
-  title: "White List",
-  description: "Generated by create next app",
-};
+// export const metadata = {
+//   title: "White List",
+//   description: "Generated by create next app",
+// };
 
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <WagmiConfig config={wagmiConfig}>
+          <Header />
+          {children}
+        </WagmiConfig>
+
+        <Web3Modal projectId={projectId} ethereumClient={ethereumClient} />
+      </body>
     </html>
   );
 }
