+++ frontend/src/components/header.js
@@ -1,0 +1,6 @@
+"use client";
+import { Web3Button } from "@web3modal/react";
+
+export default function Header() {
+  return <Web3Button />;
+}
